Fix: Limit sell logic for quantity deduction

+++ services/order/src/services/orderService.ts
@@ -105,39 +105,50 @@ const service = {
       throw new Error("Error saving limit sell transaction into database");
     }
 
-    const result = await matEngSvc.placeLimitSellOrder(limitSellRequest);
+    let ownedStock: StockOwned | null;
+    try {
+      ownedStock = await stockOwnedRepository
+        .search()
+        .where("stock_id")
+        .equals(stock_id)
+        .and("user_name")
+        .equals(user_name)
+        .returnFirst();
+      if (!ownedStock) throw new Error("User does not own this stock (placeLimitSellOrder)");
+    } catch (err) {
+      throw new Error("Error fetching owned stock data from database");
+    }
 
-    if (result.success) {
-      let ownedStock: StockOwned | null;
+    if (ownedStock.current_quantity < quantity) {
+      const transactionEntityId = transaction[EntityId];
+      if (transactionEntityId) await stockTransactionRepository.remove(transactionEntityId);
+      throw new Error(
+        `Insufficient shares. You currently own ${ownedStock.current_quantity} shares, but attempted to sell ${quantity} shares. (placeLimitSellOrder)`
+      );
+    }
+
+    const limitSellResponse = await matEngSvc.placeLimitSellOrder(limitSellRequest);
+
+    if (!limitSellResponse.success) {
+      throw new Error("Failed to place limit sell order through matching engine.");
+    }
+
+    if (ownedStock.current_quantity - quantity === 0) {
+      const ownedStockEntityId = ownedStock[EntityId];
+
+      // If the quantity to sell equals the owned quantity, delete the record
       try {
-        ownedStock = await stockOwnedRepository
-          .search()
-          .where("stock_id")
-          .equals(stock_id)
-          .and("user_name")
-          .equals(user_name)
-          .returnFirst();
-        if (!ownedStock) throw new Error("User does not own this stock (placeLimitSellOrder)");
+        if (ownedStockEntityId) await stockOwnedRepository.remove(ownedStockEntityId);
       } catch (err) {
-        throw new Error("Error fetching owned stock data from database");
+        throw new Error("Error deleting owned stock record from database");
       }
-
-      if (ownedStock.current_quantity <= quantity) {
-        // If the quantity to sell equals the owned quantity, delete the record
-        try {
-          const ownedStockEntityId = (ownedStock as any)[EntityId]; // HACK: Get's the Entity id of ownedStock and bypasses ts typecheck
-          await stockOwnedRepository.remove(ownedStockEntityId);
-        } catch (err) {
-          throw new Error("Error deleting owned stock record from database");
-        }
-      } else {
-        // If there's more stock left, just decrease the quantity
-        ownedStock = { ...ownedStock, current_quantity: ownedStock.current_quantity - quantity };
-        try {
-          await stockOwnedRepository.save(ownedStock);
-        } catch (err) {
-          throw new Error("Error updating owned stock quantity in database");
-        }
+    } else {
+      // If there's more stock left, just decrease the quantity
+      ownedStock = { ...ownedStock, current_quantity: ownedStock.current_quantity - quantity };
+      try {
+        await stockOwnedRepository.save(ownedStock);
+      } catch (err) {
+        throw new Error("Error updating owned stock quantity in database");
       }
     }
   },
